Clean up test folder structure
