+++ test/src/apputils/vdom.spec.ts
