+++ test/src/coreutils/activitymonitor.spec.ts
