+++ test/src/coreutils/observablemap.spec.ts
