+++ test/src/index.ts
@@ -7,6 +7,13 @@ import './application/instancetracker.spec';
 import './application/loader.spec';
 import './application/shell.spec';
 
+import './apputils/dialog.spec';
+import './apputils/iframe.spec';
+import './apputils/jsoneditor.spec';
+import './apputils/sanitizer.spec';
+import './apputils/styling.spec';
+import './apputils/vdom.spec';
+
 import './cells/model.spec';
 import './cells/widget.spec';
 
@@ -18,20 +25,6 @@ import './codemirror/factory.spec';
 
 import './commandlinker/commandlinker.spec';
 
-import './coreutils.spec';
-import './apputils.spec';
-import './apputils.spec';
-import './coreutils.spec';
-import './coreutils.spec';
-import './coreutils.spec';
-import './coreutils.spec';
-import './common/metadata.spec';
-import './apputils.spec';
-import './apputils.spec';
-import './coreutils.spec';
-import './coreutils.spec';
-import './apputils.spec';
-
 import './completer/handler.spec';
 import './completer/model.spec';
 import './completer/widget.spec';
@@ -40,6 +33,14 @@ import './console/foreign.spec';
 import './console/history.spec';
 import './console/panel.spec';
 import './console/widget.spec';
+
+import './coreutils/activitymonitor.spec';
+import './coreutils/observablejson.spec';
+import './coreutils/observablemap.spec';
+import './coreutils/observablestring.spec';
+import './coreutils/observablevector.spec';
+import './coreutils/time.spec';
+import './coreutils/undoablevector.spec';
 
 import './csvwidget/table.spec';
 import './csvwidget/toolbar.spec';
